MESA FIX : PART 1 (Fixing USE and adding -9999)
== Fixing USE ==
Description:
The current USE for the graphic chain for the installation of a unwanted driver, including useless dependencies because of the useless drivers.
the libdrm ebuild have the VIDEO_CARDS "radeon" to enable radeon drm for all cards (from r100 to r800).
the mesa ebuild have the VIDEO_CARDS "radeon" for Radeon r100 to r500 inclusivement. Also, he have VIDEO_CARDS "r600" for card from r600 to r800. The VIDEO_CARDS "r300" do the same thing as "radeon".

Affected:
Radeon card with r600 chip or better. Owner of Radeon r600+ need to put VIDEO_CARDS="radeon r600", enabling the radeon DRM and forcing the installation of the r300 MESA drivers even if it's not needed. The r300 MESA drivers have llvm in the dependencie and that's not used in the r600 driver.

fix:
libdrm now have the VIDEO_CARD : radeon, r300 and r600. so any of them will install the radeon drm. No need to put VIDEO_CARDS="radeon r600" for r600+ owner
mesa will not put the radeon use for r300 or r600, it just accept any of these for his libdrm dependencie.
xf86-video-ati will not put the radeon use for r300 or r600, it just accept any of these for his libdrm dependencie.
xorg-drivers will not put the radeon use for r300 or r600, it just accept any of these for his libdrm dependencie.

== -9999 ==
I have added the experimental git version of libdrm, mesa, and xf86-video-ati. These are usually better the the "stable" one. Faster less buggy and more supported ClosedGL commands.